fix service startup issue due to incorrect group permissions
Remove the hack "fix" for rhbz#1019487 as it causes
openstack-neutron-openvswtich to be installed before
the openstack-neutron package, and thus before the
neutron group is in place, thus causing invalid permissions
on the config files and ultimately service startup failure.
The hack is too big of a hammer in any case.

Fixes rhbz#1080650