Use configuration directory to configure L3 agent
There is now no guarantee that a stable list of configuration files is
used with L3 agent, so instead of listing each configuration file in the
arguments passed to the service, create and populate a directory.

The idea is to have just symlinks in the directory.

There are similar patches for devstack and grenade in u/s:

- grenade: https://review.openstack.org/163773
- devstack: https://review.openstack.org/163490

Change-Id: Ib8cb3c508386a58552309e242803f543f177896b