Introduce user defined configuration mechanism for services
If a user wants to add a new configuration file for a service to read,
he now just drops it into appropriate directory under:

/etc/neutron/conf.d/<service-name>

README file that describes this mechanism is added in the conf.d
directory.

Common files, including the directory, are moved into a new subpackage
(openstack-neutron-common), and it's now made a dependency for all
packages that ship service files.

The new package depends on python-neutron so that we don't need to add
two dependencies to each service subpackage (one for
openstack-neutron-common, another one for python-neutron).

Note: rootwrap may be of no use on API node, so if/when we decide to
split openstack-neutron subpackage into separate per-service pieces, we
may need to move rootwrap files that are now located in
openstack-neutron-common, into another subpackage that would be
installed only for those services that actually need rootwrap bits
(agents). But for now, it's not needed.

Resolves rhbz: 1188480
Change-Id: I731f9d2609b78448370bec2c7a634b15e974bef3