Enable rootwrap daemon by default
Upstream gate already tests against the daemon since Kilo. The only
reason why it's not the default value is that default root_helper is
also not rootwrap but plain 'sudo'. Since we use rootwrap, there is no
reason not to use the daemon.

Note that if someone switched back to 'sudo', then they may need to
disable the option in their neutron.conf to avoid using rootwrap in
cases where daemon would be usually considered applicable.

In that case, they would need to resolve '.rpmnew' conflicts anyway, so
a package update should not result in rootwrap daemon enablement on its
own.

Change-Id: I8fc9db2edbc1a2575c1a88445d63cc3d59fd47af