Don't pretend /etc/sudoers.d/neutron is a configuration file
It should not be handled as another configuration file, since neutron
agents depend on those particular entries being in the file, and there
seems to be no clear cases to introduce anything custom here.

If users want to apply some local configuration changes to sudoers, they
are always welcome to create another file in that same directory for
sudo consumption.

There is a slight chance to break someone who, for unknown reason,
applied irrelevant configuration changes in that file. The package
update would then drop his local changes completely. Note it's not a big
difference from now anyway since the file is not marked with
'noreplace', meaning that old contents just go into .rpmsave, and new
ones are automatically applied.

Change-Id: I3395910e66d70d0d140fdadfad75a147365b676a