Handle new ovs/linuxbridge agent configuration files
...while supporting upgrade from systems that used old configuration
file.

To achieve this, we imitate noreplace behaviour in %post for
-linuxbridge and -openvswitch packages if it's an upgrade for a system
that has the old configuration file in place. Specifically, we copy the
default new configuration file into .rpmnew and then move the old file
into the new location. Then users can manually merge the two new files,
as they do for other config file updates.

While at it, added missing %post sections for all subpackages with a
service file.

The config files were moved in upstream as part of:

- https://review.openstack.org/#/c/195277/

Change-Id: I4ca05abe75d17b43474e1869d6fec9189f96b991