Second attempt to introduce rpm version into neutron-server output
The previous one was reverted [1] because it did not work for
version-release strings used by delorean.

The new version should work for both RDO and delorean versioning
schemas.

We need to cut of %dist from %release before passing it into PBR_VERSION
because pbr assumes a version subsection with letters means a
pre-release that should conform to specific format (f.e. a, b, or dev
prefixes are allowed, but not fc or el).

[1]: I669a671dafe92cb19787dc8a7e826d50135b59f7

Change-Id: Icb6dee2665f75c659e5b5fdc06716607b850b925