openstack-neutron-common should require shadow-utils
The user 'neutron' is created as part of the %pre section of
openstack-neutron-common installation process.
Currently, shadow-utils is a requirement of openstack-neutron which is
not a dependency of openstack-neutron-common.
Hence, if openstack-neutron is not installed before hand the user
creation may fail.

Conflicts:
	openstack-neutron.spec

Resolves rhbz: 1322522

Change-Id: Id8e15158fd60ee735e2cea00f5845b478870b331
(cherry picked from commit 063c131956073d261f1982270c99a4f283fd32ec)