Load br_netfilter if present for iptables aware agents
If iptables firewall is used for ovs or linuxbridge agent, and if a
newer kernel (3.18+) is used, then we should specifically load
br_netfilter module before starting those agents, because otherwise
firewall may not work for bridged traffic (meaning, security groups for
instances may not work).

We achieve that behaviour using ExecStartPre= systemd directive. Since
we want RDO to work with older kernels as well, the command to execute
always succeeds, even if the kernel module is not present (pre 3.18).
Since we execute agents as 'neutron' user while those commands we need
to set up the bridge (modprobe, sysctl) require 'root' permissions, we
also add PermissionsStartOnly=true for those unit files.

An alternative for the fix would be to make the iptables driver itself
to load the kernel module, something that may not be in the best
interest of Neutron itself.

Another alternative would be to load the kernel module on system boot
via modules-load.d(5) subsystem. That would require another system
restart though. It would also mean exposing the kernel loading through a
configuration file, with all its fragility.

More details in: https://bugs.launchpad.net/neutron/+bug/1622914

Change-Id: I2f810262b50901e8447e7f48be1de414850ce0be