Seperate entrypoint for tempest plugin
Previously python-neutron package installed the entry point for the in-tree
Tempest plugin which cause failure while running tempest list-plugins
because Tempest found the entry point but does not have Tempest plugin
code which is subpackaged in python-neutron-tests.
Separate neutron_tests egg-info is created with the plugin entry point
so that Tempest includes the plugin only when subpackage is installed.

Change-Id: Ie294807284e64c9572cc42e1129af1d510ba1eb9