Update patch set 7
Patch Set 7:

> Where does this patch fit considering https://review.rdoproject.org/r/#/c/3436/
 > ? Does the code belong to neutron-netns-cleanup tool itself (in
 > neutron upstream)?

There's already a bug open for this:
https://bugs.launchpad.net/neutron/+bug/1403455

Checked with Assaf at the beginning and he thought that the approach in this patch would still be valid since the script could be run manually. However, I could take over the bug above and try to do the cleanup using the approach here (in python). What do you guys think?
Thanks

Patch-set: 7