Update patch set 7
Patch Set 7:

Daniel, I was thinking that the script is triggered in our usual HA flow. But if not, then maybe we indeed should kill it as Jakub suggested.

On general stance, how does neutron-netns-cleanup fit in new HA architecture? Do we call it there at all, or do we fence/cleanup a failed node in some other way? More generally, which issue do we exactly try to solve here?

Patch-set: 7