Update patch set 3
Patch Set 3:

I am for cleaning up the script. We could probably wait until we implement the missing cleanup steps in the neutron tool itself, so that operators that may still utilize the script have something to switch to; but it may take a while to polish a neutron fix, and maybe it's better to nag operators interested in the functionality by removing the script at the start of the cycle to indicate that we are not going to ship it in Ocata final.

Patch-set: 3