Update patch set 1
Patch Set 1:

> or make os.xenapi import optional upstream?

I think this is going the same path as os-win. Once we have the platform code in an OpenStack library, everyone will use it and it's hard to avoid it without creating a large distro-only patch. We're doing a similar hack in Nova, and we could avoid it with a packaged os-xenapi.

Patch-set: 1