Load br_netfilter on system load
Calling the first 'exec' in the script replaces the current process, so
the second 'exec' statement is never reached. It means that
br_netfilter module is not loaded early in system startup.

This module is needed to implement iptables based firewall in Neutron,
and without it, security groups are not effective.

Good news is this module is also loaded on L2 agent startup. In the
future, we may want to remove module loading from L2 agent systemd
service ExecPre= directive, at which point it would be an issue.

Change-Id: I39f43bc7b15ced187573124eae95b1e26d0fbb4c
(cherry picked from commit cd7ce6e318cd36c669d2bb6d415b8716408926fe)