Update patch set 1
Patch Set 1: -Workflow

ah sorry, (cherry picked from ...) was missing, this is indeed a backport
re. merged or not: OpenStack upstream stable-maint policy requires to be merged on the master branch first
This should not be a problem here, just that we need to update neutron-distgit cores since some folks moved on.

Patch-set: 1