Update patch set 19
Patch Set 19:

Note that once we merge this patch, we will break tripleo jobs with octavia and ovn backend as the tripleo containers builds installs python3-networking-ovn in the container ant after patch that is python3-neutron which doesn't provide the plugin:

https://opendev.org/openstack/tripleo-common/src/branch/master/container-images/tripleo_kolla_template_overrides.j2#L335

We need a new package for ovn-octavia-provider and fix that asap.

As workaround and if the content of networking-ovn is still valid for octavia, we may temporarly add a Provides: python3-ovn-octavia-provider to the current package anf push a fix for tripleo-common to move to it.

Patch-set: 19
Label: Code-Review=0