Sync neutron dependencies with upstream requirements
This patch removes python-os-xenapi from the package dependencies
as it was dropped from upstream neutron some time ago.
It also adds python-packaging to the dependencies as it was added
in upstream recently.
It also bumps versions of some other packages to be aligned with
what is in the upstream neutron's requirements.

Change-Id: I3d943414c3a2ffd5a6185954da5d09282712cad6