Add privsep-helper to neutron sudoers file
Neutron uses the oslo.privsep library which provides the privsep-helper
command to privilege escalations. This command should be runnnable by
the neutron user via sudo without a password.

The current sudoers file does not allow the privsep-helper and users
always need to use the privsep-helper command via the rootwrap command.

Change-Id: I21b3c790d217cc2d52a05652f615b7aa32f1783f