Remove testrepository dependency
In 2018 the neutron project switched from testrepository to stestr.
Hence this dependency is not needed anymore and for OSP 18 where
the python3-testrepository RPM is not in main repo it causes troubles.

The python3-os-testr which uses python3-stestr is already listed in spec.

Related: https://review.opendev.org/c/openstack/neutron/+/566283
Change-Id: I5373a9f392f74e2993983c92c07e8b89d410032b